rename SysId example to SysIdRoutine
CMake target output conflicts with sysid (the application) on windows
