+++ wpilibjExamples/src/main/java/edu/wpi/first/wpilibj/examples/examples.json
@@ -857,7 +857,7 @@
       "Command-based",
       "DataLog"
     ],
-    "foldername": "sysid",
+    "foldername": "sysidroutine",
     "gradlebase": "java",
     "commandversion": 2,
     "mainclass": "Main"
