+++ wpilibjExamples/src/main/java/edu/wpi/first/wpilibj/examples/sysidroutine/Robot.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package edu.wpi.first.wpilibj.examples.sysid;
+package edu.wpi.first.wpilibj.examples.sysidroutine;
 
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj2.command.Command;
