+++ wpilibjExamples/src/main/java/edu/wpi/first/wpilibj/examples/sysidroutine/SysIdRoutineBot.java
@@ -2,11 +2,11 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package edu.wpi.first.wpilibj.examples.sysid;
+package edu.wpi.first.wpilibj.examples.sysidroutine;
 
-import edu.wpi.first.wpilibj.examples.sysid.Constants.OIConstants;
-import edu.wpi.first.wpilibj.examples.sysid.subsystems.Drive;
-import edu.wpi.first.wpilibj.examples.sysid.subsystems.Shooter;
+import edu.wpi.first.wpilibj.examples.sysidroutine.Constants.OIConstants;
+import edu.wpi.first.wpilibj.examples.sysidroutine.subsystems.Drive;
+import edu.wpi.first.wpilibj.examples.sysidroutine.subsystems.Shooter;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
